+++ RELEASE-NOTES.md
@@ -1,6 +1,13 @@
 OpenService-BCB
 ===============
 
+0.1.2 - 2017-01-16
+------------------
+
+- Improvement: secondary index on table `stats_bcb_history` is no longer needed.
+- New application config `rankings_update_batch_mode: true/false`
+
+
 0.1.1 - 2017-01-15
 ------------------